Fix issue with GET requests (#79) (#85)
The haystack operations about, ops, and formats are all typically sent
as GET requests to a haystack server.

Previously, these all called 'CallAsync' with an empty HaystackGrid. In
'CallAsync', it called 'PostStringAsync'. When trying to write out the
empty grid, an exception occurred because no columns were defined.

This commit updates the logic so that if the grid passed to 'CallAsync'
is null or has no columns, that it assumes the operation is a simple GET
request and responds appropriately.